feat: implement large widget
