+++ SNUTT-2022/SNUTT/Views/Components/Timetable/TimetableBlockLink.swift
@@ -1,0 +1,8 @@
+//
+//  TimetableBlockLink.swift
+//  SNUTT
+//
+//  Created by 박신홍 on 2022/07/23.
+//
+
+import Foundation
